tests: move recon tests according to the class they actually test
Moves two test cases implemented in StandardReconConfigTests which
are actually testing ReconOperation to ReconOperationTests.
